Remove old threshold tests and retain only the pytest versions. (#1991)
